% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_stepwise_regression.R
\name{AIC_stepwise_regression}
\alias{AIC_stepwise_regression}
\title{Logistic regression using stepwise Akaike information criteria selection}
\usage{
AIC_stepwise_regression(outcome, dataset, variables, stepwise_direction)
}
\arguments{
\item{outcome}{Outcome of interest}

\item{dataset}{Dataframe that contains the trait and exposure variables}

\item{variables}{Exposure variables of interest. Must be numeric or one-hot encoded}

\item{stepwise_direction}{Direction of stepwise selection. Options include: 'both', 'backward', or 'forward'. For more information, see stats::step.}
}
\value{
Regression results
}
\description{
Performs Akaike information criterion (AIC) informed selection using stats::step function.
}
