% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{asr}
\alias{asr}
\title{Ancestral state reconstruction wrapper function}
\usage{
asr(
  df,
  tr,
  tip_name_variable,
  trait,
  model = "ER",
  node_states = "joint",
  upper_bound = 1e+50,
  lower_bound = 1e-09,
  confidence_threshold = NULL
)
}
\arguments{
\item{df}{Dataframe with tip name (e.g., tip_name_variable) and phenotype/trait (e.g., trait) variables}

\item{tr}{Phylogeny of class "phylo"}

\item{tip_name_variable}{Name of variable containing tip names in df. Tip name variable must correspond to tip names in the tree.}

\item{trait}{Name of phenotype/trait variable in df}

\item{model}{Type of rate transition matrix. Options: equal rates model ("ER") or all rates different ("ARD"). An additional option, "MF", selects the best performing model using sample-size corrected Akaike information criterium (AICc). Default: ER}

\item{node_states}{Perform "joint" or "marginal" reconstruction. Default: joint}

\item{upper_bound}{Upper bound for likelihood search. Default: 1e50}

\item{lower_bound}{Lower bound for likelihood search. Default: 1e-9}

\item{confidence_threshold}{The confidence threshold to use when categorizing ancestral state inferences from marginal reconstruction. We highly suggest using a default of 0.5 (i.e., winner-takes-all), but permit modification to elevated values.}
}
\value{
\describe{
    \item{corHMM_output}{corHMM output}
    \item{corHMM_model_statistics}{A dataframe containing the inferred rates, log-likelihood, AIC, and chosen transition model}
    \item{parent_child_df}{A dataframe of parent-child relationships with additional descriptive information}
    \item{node_states}{Text string indicating the chosen reconstruction method}
  }
}
\description{
Function to perform ancestral state reconstruction using corHMM. Output is parsed into a dataset with episodes of trait gain, loss, and continuation across the phylogeny
}
