% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_clustering_algorithm.R
\name{asr_clustering_algorithm}
\alias{asr_clustering_algorithm}
\title{Generate Ancestral Reconstruction and Clustering Data}
\usage{
asr_clustering_algorithm(
  df,
  tr,
  pheno,
  type = "discrete",
  model = "ER",
  conf_threshold = 0.875,
  remove_faux = "yes",
  confidence = "high",
  remove_revertant = "yes",
  collapse_cluster = "yes"
)
}
\arguments{
\item{df}{A dataframe containing isolate numbers and patient IDs columns labelled as `$isolate_no` and `$Patient_ID`, respectively.}

\item{tr}{A phylogenetic tree object.}

\item{pheno}{A string specifying the column name in `df` representing the binary phenotype to be analyzed.}

\item{type}{A character string specifying the type of ancestral reconstruction model (e.g., "discrete").}

\item{model}{A character string specifying the model used for ancestral state reconstruction (e.g., "ER" for equal rates).}

\item{conf_threshold}{A numeric value specifying the confidence threshold for filtering parent-child relationships.}

\item{remove_faux}{A character string ("yes" or "no") indicating whether to remove or relabel faux clusters.}

\item{confidence}{A character string ("high or "low") specifying the confidence level for identifying the transitions in the clustering data.}

\item{remove_revertant}{A character string ("yes" or "no") indicating whether to remove revertant calls from the cleaned text string}

\item{collapse_cluster}{A character string ("yes or "no") indicating whether to create an additional variable where all singletons and cluster isolates (i.e., cluster 1, cluster 2) are collapsed into one a general category (i.e., singleton OR cluster).}
}
\value{
A list containing the following elements:
  \describe{
    \item{asr_model}{A dataframe of the ancestral state reconstruction results.}
    \item{asr_model_statistics}{A dataframe with statistics on the ancestral state reconstruction model}
    \item{parent_child_df}{A dataframe of parent-child relationships with additional descriptive information.}
    \item{tip_data_df}{A dataframe of tip data merged with clustering information.}
  }
}
\description{
This function performs ancestral state reconstruction on a given phylogenetic tree and
phenotypic data, generates parent-child relationships, and identifies clusters.
It returns a list containing the ancestral reconstruction model, parent-child
relationships, and tip data merged with clustering information.
}
\examples{
\dontrun{
df <- phylosuite::phylosuiteR_data
tr <- phylosuite::tr
result <- asr_clustering_algorithm(df = df, tr = tr, pheno = "MVB_non_s",
    type = "discrete", model = "ER", conf_threshold = 0.875,
    remove_faux = "yes", confidence = "high")
}
}
