% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_asr_model.R
\name{characterize_asr_model}
\alias{characterize_asr_model}
\title{Characterize ancestral state reconstruction model}
\usage{
characterize_asr_model(corHMM_obj)
}
\arguments{
\item{corHMM_obj}{Ancestral reconstruction model from corHMM::corHMM}
}
\value{
Dataframe containing model fit (i.e., log-likelihood and AIC), rates, and model type (i.e., ER or ARD)#'   \describe{
    \item{model}{Model}
    \item{number_parameters}{Number of parameters in model}
    \item{number_rate_categories}{Number of rate categories}
    \item{rate1}{Rate 1}
    \item{rate2}{Rate 2}
    \item{loglik}{Log-likelihood of model}
    \item{AIC}{Akaike information criteria}
    \item{AICc}{Sample size corected Akaike information criteria}
  }
}
\description{
Function to report model statistics for a binary ancestral state reconstruction model
}
