% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Dataframe with relevant data}
\format{
## `df`
Dataframe with 413 rows and 4 columns
}
\source{
<https://github.com/Snitkin-Lab-Umich/ltach-crkp-colistin-ms>
}
\usage{
df
}
\description{
413 carbapenem-resistant Klebsiella pneumoniae isolates from long-term acute care hospitals across the United States of America. Whole-genome sequencing and antimicrobial susceptibility testing were performed on each isolate
}
\keyword{datasets}
