% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_gain_loss.R
\name{downstream_gain_loss}
\alias{downstream_gain_loss}
\title{downstream_gain_loss: Analysis of downstream gain and loss of comparitor traits}
\usage{
downstream_gain_loss(
  comparitor_parent_child_df,
  trait_parent_child_df,
  tr,
  node_states = "joint",
  confidence = NULL
)
}
\arguments{
\item{comparitor_parent_child_df}{Parent child dataset for a comparitor trait, such as a genotype}

\item{trait_parent_child_df}{Parent child dataset for a trait of interest, such as a phenotype}

\item{tr}{Phylogenetic tree}

\item{node_states}{Joint or marginal reconstruction}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Summary stats for downstream gain and loss of a trait
}
\description{
Downstream gain and loss of a comparitor trait on stretches of a different trait of interest
}
