% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_permutation_test.R
\name{downstream_permutation_test}
\alias{downstream_permutation_test}
\title{downstream_permutation_test: Permutation test for downstream analysis}
\usage{
downstream_permutation_test(
  genotype,
  df,
  tr,
  tip_name_var,
  pheno,
  node_states = "joint",
  num_permutations = 1000,
  num_cores = 6
)
}
\arguments{
\item{tr}{Phylogenetic tree}

\item{node_states}{Joint or marginal reconstruction}

\item{comparitor_parent_child_df}{Parent child dataset for a comparitor trait, such as a genotype}

\item{trait_parent_child_df}{Parent child dataset for a trait of interest, such as a phenotype}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Summary stats for downstream gain and loss of a trait
}
\description{
Permutation test for analysis of downstream gain and loss of a comparitor trait on stretches of a different trait of interest
}
