% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_transitions.R
\name{downstream_transitions}
\alias{downstream_transitions}
\title{Analysis of downstream transitions of comparitor traits}
\usage{
downstream_transitions(
  comparitor_parent_child_df,
  trait_parent_child_df,
  tr,
  node_states = "joint",
  confidence = NULL
)
}
\arguments{
\item{comparitor_parent_child_df}{Parent child dataset for a comparitor trait, such as a genotype or a different trait/phenotype}

\item{trait_parent_child_df}{Parent child dataset for a trait of interest, such as a trait/phenotype}

\item{tr}{Phylogenetic tree}

\item{node_states}{Joint or marginal reconstruction}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Summary stats for downstream gain and loss of a trait
}
\description{
Downstream gain and loss of a comparitor trait on stretches of a different trait of interest
}
