% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_asr_model.R
\name{find_best_asr_model}
\alias{find_best_asr_model}
\title{find_best_asr_model: Determine best model for ancestral state reconstruction}
\usage{
find_best_asr_model(
  df,
  tr,
  tip_name_variable,
  trait,
  node_states = "joint",
  upper_bound = 1e+50,
  lower_bound = 1e-09
)
}
\arguments{
\item{df}{Dataframe with tip name (e.g., tip_name_variable) and phenotype/trait (e.g., trait) variables}

\item{tr}{Phylogenetic tree}

\item{tip_name_variable}{Name of variable containing tip names in df}

\item{trait}{Name of phenotype/trait variable in df}

\item{node_states}{Whether to perform "joint" or "marginal" reconstruction}

\item{upper_bound}{Upper bound for likelihood search. Default: 1e50}

\item{lower_bound}{Lower bound for likelihood search. Default: 1e-9}
}
\value{
Description of return value
  \describe{
    \item{model_statistics}{Summary of model options}
    \item{best_model}{Best model for ancestral state reconstruction}
    \item{node_states}{Whether "joint" or "marginal" reconstruction was performed}
    \item{corHMM_output_best_model}{corHMM output for best model}
  }
}
\description{
Function to find best model for ancestral state reconstruction using corHMM.
}
