% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genotype_mat}
\alias{genotype_mat}
\title{Genotype matrix}
\format{
## `genotype_mat`
Dataframe with 413 rows and 76 columns
}
\source{
<https://github.com/Snitkin-Lab-Umich/ltach-crkp-colistin-ms>
}
\usage{
genotype_mat
}
\description{
Row names correspond to isolate names, as found in the tr$tip.name and df$tip_name_var. Columns correspond to colistin-related genotypes identified in the aforementioned study. These genotypes can be leveraged to test for synchronous and downstream phenotype-genotype transition events with the synchronous_detection() and downstream_gain_loss() functions, respectively
}
\keyword{datasets}
