% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{get_continuation_data}
\alias{get_continuation_data}
\title{get_continuation_data: Annotate parent child data from ancestral state reconstructionwith transition data}
\usage{
get_continuation_data(parent_child_df, node_states)
}
\arguments{
\item{parent_child_df}{Edge matrix with ancestral states}

\item{node_states}{Whether to perform "joint" or "marginal" reconstruction. Default: joint}
}
\value{
Description of return value
  \describe{
    \item{parent_child_df}{Final parent child dataset with transition data}
  }
}
\description{
Function to generate an annotated edge matrix with transition data
}
