% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{get_parent_child_data}
\alias{get_parent_child_data}
\title{get_parent_child_data: Get parent child data from ancestral state reconstruction}
\usage{
get_parent_child_data(
  tr,
  anc_data,
  pheno_data,
  conf_threshold = NULL,
  node_states
)
}
\arguments{
\item{tr}{Phylogenetic tree object. Class must be phylo}

\item{anc_data}{Ancestral states inferred by corHMM}

\item{pheno_data}{Named phenotype string}

\item{conf_threshold}{The confidence threshold to use for marginal state reconstruction. Suggested value: 0.875.}

\item{node_states}{Whether to perform "joint" or "marginal" reconstruction. Default: joint}
}
\value{
Description of return value
  \describe{
    \item{edge}{Edge matrix with states}
  }
}
\description{
Function to generate an annotated edge matrix
}
