% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_algorithm.R
\name{nearest_neighbor_algorithm}
\alias{nearest_neighbor_algorithm}
\title{Identify the phylogenetic nearest neighbor of a tip}
\usage{
nearest_neighbor_algorithm(
  isolate,
  tip_name_variable,
  phylogenetic_distance,
  variant_distance,
  metadata,
  annotate = NULL,
  variables_of_interest = NULL,
  comparison = NULL,
  comparison_feature = NULL
)
}
\arguments{
\item{isolate}{Isolate of interest}

\item{tip_name_variable}{Name of variable containing tip names. Must be found in the metadata, dataframe. Tip name entries must also correspond to tip names in the tree}

\item{phylogenetic_distance}{Matrix of phylogenetic distance. Generated by ape::cophenetic.phylo(tree)}

\item{variant_distance}{(OPTIONAL) Matrix with secondary distance measure, presumably SNPs. If not specified, the first entry will be gathered}

\item{metadata}{Dataframe with trait and variables of interest}

\item{annotate}{Boolean (TRUE/FALSE) indicating if annotation is necessary}

\item{variables_of_interest}{Variables found in the metadata}

\item{comparison}{Boolean (TRUE/FALSE) indicating whether to restrict analysis to isolates with/without a trait (comparison_feature)}

\item{comparison_feature}{Name of variable to compare}
}
\value{
Dataframe with the nearest neighbor of interest.
}
\description{
This function uses phylogenetics to identify an isolate's nearest neighbor.
}
