% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_summary_statistics.R
\name{nearest_neighbor_summary_statistics}
\alias{nearest_neighbor_summary_statistics}
\title{nearest_neighbor_summary_statistics: Generate summary statistics of nearest neighbor comparisons for isolate pairs}
\usage{
nearest_neighbor_summary_statistics(
  nn_diff_df,
  categorical_vars = NULL,
  binary_vars = NULL,
  continuous_vars = NULL,
  log_2 = FALSE
)
}
\arguments{
\item{nn_diff_df}{Dataframe with isolates and their nearest neighbors, generated from nearest_neighbor_analysis.}

\item{binary_vars}{Binary variables. Must be in the comparison_df AND coded 0 and 1.}

\item{continuous_vars}{Continuous variables. Must be in the comparison_df}

\item{log_2}{Boolean (TRUE/FALSE) Whether to perform log-2 transformation on continuous variables}
}
\value{
Dataframe with differences in the categorical and continuous variables between an isolate and their nearest neighbor
}
\description{
This function will generate summary statistics for nearest neighbor comparisons
}
