% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloaware_dataset_curation.R
\name{phyloaware_dataset_curation}
\alias{phyloaware_dataset_curation}
\title{Dataset curation for phylogenetically aware regression}
\usage{
phyloaware_dataset_curation(
  trait,
  df,
  first_present = FALSE,
  patient_id = NULL,
  culture_date = NULL
)
}
\arguments{
\item{trait}{Outcome of interest. Character string.}

\item{df}{Dataset with trait variable and asr_cluster. Must contain patient_id and culture_date variables if first_present == TRUE.}

\item{first_present}{Boolean (i.e., TRUE/FALSE) indicating whether to take the first present isolate}

\item{patient_id}{Variable that contains IDs that group tips (i.e., Patient ID). Character string.}

\item{culture_date}{Variable that contains collection/culture date. Character string.}
}
\value{
Three datasets labeled as 'present', 'singleton', and 'cluster'.
}
\description{
Curate three datasetes for phylogenetically aware regression.
}
\details{
Datasets included:
1. Present: All isolates
2. Singleton: Singleton isolates-only
3. Cluster: Cluster isolates-only

This permits isolated analysis of singleton and cluster isolates, enabling identification of characteristics associated with the emergence and/or spread of a trait.
}
