% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloaware_regression.R
\name{phyloaware_regression}
\alias{phyloaware_regression}
\title{phyloaware_regression: Phylogenetically-aware regression}
\usage{
phyloaware_regression(
  pheno,
  variables,
  df,
  first_present = NULL,
  patient_id = NULL,
  culture_date = NULL,
  multivariable = NULL,
  stepwise_direction = NULL,
  entry_criteria = NULL,
  retention_criteria = NULL,
  confounding_criteria = NULL
)
}
\description{
This function performs regression on our dataset
}
