% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purposeful_selection_algorithm.R
\name{purposeful_selection_algorithm}
\alias{purposeful_selection_algorithm}
\title{Logistic regression using purposeful selection}
\usage{
purposeful_selection_algorithm(
  outcome,
  variables,
  dataset,
  entry_criteria = 0.2,
  retention_criteria = 0.1,
  confounding_criteria = 0.2
)
}
\arguments{
\item{outcome}{Outcome of interest}

\item{variables}{Exposure variables of interest. Must be numeric or one-hot encoded}

\item{dataset}{Dataframe that contains the trait and exposure variables}

\item{entry_criteria}{P-value criteria for entry into the model. Default = 0.2}

\item{retention_criteria}{P-value criteria for retention into the model. Default = 0.1}

\item{confounding_criteria}{Percent change of effect size. Set value to be very high (i.e., 1000) if testing for confounding is not desired. Default = 0.2.}
}
\value{
Regression results from each step.
}
\description{
Uses purposeful selection algorithm to identify a regression model.
Three steps exist:
1. Unadjusted logistic regression to identify candidate variables under a p-value threshold (entry_criteria)
2. Multivariable regression of candidate variables. Iterative process, starting from the value with the highest p-value, variables are retained if they fall under category of a significant variable (< retention_criteria) or confounding (i.e., effect size +/- confounding criteria).
3. All variables that failed step 1 are introduced to the model and retained if p-value < retention criteria (retention_criteria)
Reference: https://scfbm.biomedcentral.com/articles/10.1186/1751-0473-3-17
}
