% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_informed_regression.R
\name{pvalue_informed_regression}
\alias{pvalue_informed_regression}
\title{Logistic regression using p-values to select final model}
\usage{
pvalue_informed_regression(
  outcome,
  dataset,
  variables,
  entry_criteria = 0.2,
  retention_criteria = 0.1
)
}
\arguments{
\item{outcome}{Outcome of interest}

\item{dataset}{Dataframe that contains the trait and exposure variables}

\item{variables}{Exposure variables of interest. Must be numeric or one-hot encoded}

\item{entry_criteria}{P-value criteria for entry into the model. Default = 0.2}

\item{retention_criteria}{P-value criteria for retention into the model. Default = 0.1}
}
\value{
Final regression model
}
\description{
Performs unadjusted logistic regression to identify candidate variables that fall under an p-value threshold (entry_criteria). Forward variable selection is performed to introduce variables into the model and retain if they fall within a more stringent criteria (retention_criteria).
}
