% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous_detection.R
\name{synchronous_detection}
\alias{synchronous_detection}
\title{Synrchronous detection}
\usage{
synchronous_detection(comparitor_parent_child_df, trait_parent_child_df)
}
\arguments{
\item{x}{Description of parameter `x`}

\item{y}{Description of parameter `y`}
}
\value{
Description of return value
}
\description{
Description of what the function does.
}
