% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous_transitions.R
\name{synchronous_transitions}
\alias{synchronous_transitions}
\title{Analysis of synchronous transitions of two traits}
\usage{
synchronous_transitions(
  comparitor_parent_child_df,
  trait_parent_child_df,
  node_states = "joint",
  confidence = NULL
)
}
\arguments{
\item{comparitor_parent_child_df}{Parent child dataset for a comparitor trait, such as a genotype or a different trait/phenotype}

\item{trait_parent_child_df}{Parent child dataset for a trait of interest, such as a trait/phenotype}

\item{node_states}{Joint or marginal reconstruction}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Synchronous gain and loss events of two traits
}
\description{
Synchronous gain and loss of two traits at a spot on tree
}
