% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tr}
\alias{tr}
\title{Phylogenetic tree}
\format{
## `tr`
A tree with 413 trips:
}
\source{
<https://github.com/Snitkin-Lab-Umich/ltach-crkp-colistin-ms>
}
\usage{
tr
}
\description{
A Newick phylogenetic tree of carbapenem-resistant Klebsiella pneumoniae belonging to the epidemic lineage sequence type 258, constructed using IQ-Tree and the GTR+G substitution model, with 413 tips.
}
\keyword{datasets}
