% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{asr}
\alias{asr}
\title{asr}
\usage{
asr(
  df,
  tr,
  tip_name_var,
  pheno,
  model = "ER",
  node_states = "joint",
  conf_threshold = 0.875
)
}
\arguments{
\item{x}{Description of parameter `x`}

\item{y}{Description of parameter `y`}
}
\value{
Description of return value
}
\description{
Description of what the function does.
}
