% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{asr}
\alias{asr}
\title{asr: ancestral state reconstruction wrapper function}
\usage{
asr(
  df,
  tr,
  tip_name_var,
  pheno,
  model = "ER",
  node_states = "joint",
  upper_bound = 1e+100,
  lower_bound = 1e-09,
  conf_threshold = NULL
)
}
\arguments{
\item{df}{Dataframe with tip name variable and phenotype}

\item{tr}{Phylogenetic tree}

\item{tip_name_var}{Name of variable containing tip names in df}

\item{pheno}{Name of phenotype variable in df}

\item{model}{Whether to use equal rates "ER" or all-rates differ "ARD" rate matrices.}

\item{node_states}{Whether to perform "joint" or "marginal" reconstruction}

\item{conf_threshold}{The confidence threshold to use for marginal state reconstruction}

\item{upper.bound}{Upper bound for likelihood search. The default for my implementation is 1e9}

\item{lower.bound}{Lower bound for likelihood search. The default for my implementation is 1e-9}
}
\value{
Description of return value
  \describe{
    \item{corHMM_out}{corHMM output}
    \item{parent_child_df}{A dataframe of parent-child relationships with additional descriptive information.}
  }
}
\description{
Function to perform ancestral state reconstruction using corHMM. Parse output into informative dataset with episodes of trait gain, loss, and continuation across the phylogenetic tree
}
