% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_cluster_analysis.R
\name{asr_cluster_analysis}
\alias{asr_cluster_analysis}
\title{asr_cluster_analysis: Cluster analysis}
\usage{
asr_cluster_analysis(tip_data_df)
}
\arguments{
\item{tip_data_df}{Tip data frame object from asr_cluster_detection()}
}
\value{
A dataframe with data on the phylogenetics of a trait
}
\description{
This function calculates the clustering statistics (e.g., number of singletons, clusters, and no features)
}
