% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_cluster_analysis.R
\name{asr_cluster_analysis}
\alias{asr_cluster_analysis}
\title{Analysis of trait clustering on a phylogeny}
\usage{
asr_cluster_analysis(tip_data_df)
}
\arguments{
\item{tip_data_df}{Tip data frame object from asr_cluster_detection()}
}
\value{
A dataframe with data on the phylogenetics of a trait.
}
\description{
This function calculates the clustering statistics (e.g., number of singletons, clusters, and no features).
Alongside the frequency of these events, the size of resistance and revertant clusters are reported.
Two additional statistsics are employed:
1. Phylogenetic frequency: No. clusters + singletons / total opportunities (i.e., phylogenetic events + tips without the trait)
2. Clustering frequency:  No. clusters / total phylogenetic events (i.e., singleton events + clusters)
}
