% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_cluster_detection.R
\name{asr_cluster_detection}
\alias{asr_cluster_detection}
\title{Cluster detection}
\usage{
asr_cluster_detection(
  df,
  tr,
  pheno,
  parent_child_df,
  remove_faux = "yes",
  confidence = "high",
  remove_revertant = "yes",
  collapse_cluster = "yes"
)
}
\arguments{
\item{x}{Description of parameter `x`}

\item{y}{Description of parameter `y`}
}
\value{
Description of return value
}
\description{
Description of what the function does.
}
