% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_cluster_detection.R
\name{asr_cluster_detection}
\alias{asr_cluster_detection}
\title{Identification of trait emergence and clustering of trait across the phylogeny using transition data}
\usage{
asr_cluster_detection(
  df,
  tr,
  tip_name_variable,
  patient_id = NULL,
  parent_child_df,
  node_states = "joint",
  confidence = NULL,
  simplify_faux_clusters = FALSE,
  simplify_revertant = TRUE,
  collapse_cluster = TRUE
)
}
\arguments{
\item{df}{Dataframe with tip name variable and phenotype}

\item{tr}{Phylogenetic tree}

\item{tip_name_variable}{Name of variable containing tip names in df}

\item{patient_id}{Name of variable containing patient IDs, can be combined with faux_clusters option to factor into whether a cluster should have >1 patient. (Optional)}

\item{parent_child_df}{Parent child dataframe from asr() object}

\item{node_states}{Whether the reconstruction was "joint" or "marginal"}

\item{confidence}{Whether to use 'high' (i.e., 0 -> 1) or 'low' (0 -> 0.5) confidence transitions when determining clustering with marginal ancestral state reconstruction results. If the confidence_threshold in asr() was set > 0.5, consider setting confidence as 'low'. Otherwise, set confidence as 'high'.}

\item{simplify_faux_clusters}{Booleane (i.e., TRUE/FALSE), whether to collapse faux clusters (i.e., clusters where 1 patient contributes all isolates) as singletons without distinction (Optional)}

\item{simplify_revertant}{Boolean (i.e., TRUE/FALSE). Whether to collapse revertant episodes as isolates without the trait in the cleaned text string}

\item{collapse_cluster}{Boolean (i.e., TRUE/FALSE). Whether to create a variable that collapses cluster calls into one category}
}
\value{
A tip-only dataframe with inferences on the history of these strains. Can be merged with parent_child_df from asr() if desired
}
\description{
This function identifies episodes of trait emergence and loss across a phylogenetic tree.
}
