% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_transition_analysis.R
\name{asr_transition_analysis}
\alias{asr_transition_analysis}
\title{asr_transition_analysis: Transition analysis}
\usage{
asr_transition_analysis(parent_child_df, node_states = "joint")
}
\arguments{
\item{parent_child_df}{parent_child_df object from asr() object}

\item{node_states}{Whether this was done under joint or marginal reconstruction}
}
\value{
A dataframe with data on the trait dynamics
}
\description{
This function calculates the transition statistics for ancestral state
}
