% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_transition_analysis.R
\name{asr_transition_analysis}
\alias{asr_transition_analysis}
\title{asr_transition_analysis: Transition analysis}
\usage{
asr_transition_analysis(parent_child_df, node_states = "joint")
}
\arguments{
\item{parent_child_df}{parent_child_df object returned sfrom asr() function}

\item{node_states}{If 'joint' or 'marginal' reconstruction was used}
}
\value{
Dataframe with phylogenetic transition statistics for a trait
}
\description{
Calculate transition statistics (e.g., gain, loss, and continuation) of an ancestral state reconstruction model.
}
