% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterize_asr_model.R
\name{characterize_asr_model}
\alias{characterize_asr_model}
\title{Characterize binary ancestral state reconstruction model}
\usage{
characterize_asr_model(corHMM_obj)
}
\arguments{
\item{corHMM_obj}{Ancestral reconstruction model from corHMM::corHMM}
}
\value{
A dataframe containing the log-likelihoods of the models, the rates and standard errors, and the model type
}
\description{
This function provides general model statistics for a binary ancestral state reconstruction model
}
