% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_gain_loss.R
\name{downstream_gain_loss}
\alias{downstream_gain_loss}
\title{Downstream gain loss}
\usage{
downstream_gain_loss(comparitor_parent_child_df, parent_child_df, tr)
}
\arguments{
\item{x}{Description of parameter `x`}

\item{y}{Description of parameter `y`}
}
\value{
Description of return value
}
\description{
Description of what the function does.
}
