% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_permutation_test.R
\name{downstream_permutation_test}
\alias{downstream_permutation_test}
\title{downstream_permutation_test: Permutation test for downstream analysis}
\usage{
downstream_permutation_test(
  comparitor,
  df,
  tr,
  tip_name_var,
  trait,
  node_states = "joint",
  conf_threshold = NULL,
  confidence = NULL,
  num_permutations = 1000,
  num_cores = 6
)
}
\arguments{
\item{comparitor}{Comparitor trait, such as a genotype}

\item{df}{Dataframe with trait and comparitor and tip_name_var}

\item{tr}{Phylogenetic tree}

\item{tip_name_var}{Tip name variable}

\item{node_states}{Joint or marginal reconstruction}

\item{conf_threshold}{Set value for confidence threshold of MLE calls if marginal}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Summary stats for downstream gain and loss of a trait
}
\description{
Permutation test for analysis of downstream gain and loss of a comparitor trait on stretches of a different trait of interest
}
