% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_permutation_test.R
\name{downstream_permutation_test}
\alias{downstream_permutation_test}
\title{Downstream}
\usage{
downstream_permutation_test(
  genotype,
  df,
  tr,
  tip_name_var,
  pheno,
  node_states = "joint",
  num_permutations = 1000,
  num_cores = 6
)
}
\arguments{
\item{x}{Description of parameter `x`}

\item{y}{Description of parameter `y`}
}
\value{
Description of return value
}
\description{
Description of what the function does.
}
