% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_asr_model.R
\name{find_best_asr_model}
\alias{find_best_asr_model}
\title{Determine best model for ancestral state reconstruction}
\usage{
find_best_asr_model(
  df,
  tr,
  tip_name_variable,
  trait,
  node_states = "joint",
  upper_bound = 1e+50,
  lower_bound = 1e-09
)
}
\arguments{
\item{df}{Dataframe with tip name (e.g., tip_name_variable) and phenotype/trait (e.g., trait) variables}

\item{tr}{Phylogeny object of class "phylo"}

\item{tip_name_variable}{Name of variable containing tip names in the dataframe (df). Tip name variable must correspond to tip names in the tree.}

\item{trait}{Name of phenotype/trait variable in the dataframe (df).}

\item{node_states}{Perform "joint" or "marginal" reconstruction. Default: joint}

\item{upper_bound}{Upper bound for likelihood search. Default: 1e50}

\item{lower_bound}{Lower bound for likelihood search. Default: 1e-9}
}
\value{
\describe{
    \item{model_statistics}{Dataframe with summary of model options}
    \item{best_model}{Character string indicating best model (i.e., "ER" or "ARD)}
    \item{node_states}{Characetr string whether "joint" or "marginal" reconstruction was performed}
    \item{corHMM_output_best_model}{corHMM output for best model}
  }
}
\description{
Function to find best model for ancestral state reconstruction using corHMM.
}
