% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_asr_model.R
\name{find_best_asr_model}
\alias{find_best_asr_model}
\title{find_best_asr_model: Determine best model for ancestral state reconstruction}
\usage{
find_best_asr_model(
  df,
  tr,
  tip_name_var,
  pheno,
  node_states = "joint",
  upper_bound = 1e+100,
  lower_bound = 1e-09
)
}
\arguments{
\item{df}{Dataframe with tip name variable and phenotype}

\item{tr}{Phylogenetic tree}

\item{tip_name_var}{Name of variable containing tip names in df}

\item{pheno}{Name of phenotype variable in df}

\item{node_states}{Whether to perform "joint" or "marginal" reconstruction}

\item{upper_bound}{Upper bound for likelihood search. Default: 1e100}

\item{lower_bound}{Lower bound for likelihood search. Default: 1e-9}
}
\value{
Description of return value
  \describe{
    \item{model_options}{Summary of model options}
    \item{best_model}{Best model to use for the ancestral state reconstruction}
  }
}
\description{
Function to find best model for ancestral state reconstruction using corHMM.
}
