% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{get_parent_child_data}
\alias{get_parent_child_data}
\title{Get parent child data from ancestral state reconstruction}
\usage{
get_parent_child_data(
  tr,
  ancestral_states,
  trait_data,
  node_states,
  confidence_threshold = NULL
)
}
\arguments{
\item{tr}{Phylogenetic tree of class "phylo"}

\item{ancestral_states}{Ancestral states inferred by corHMM}

\item{trait_data}{Named phenotype/trait string}

\item{node_states}{Perform "joint" or "marginal" reconstruction. Default: joint}

\item{confidence_threshold}{The confidence threshold to use for marginal state reconstruction. Suggested value: 0.875.}
}
\value{
\describe{
    \item{edge}{Edge matrix with ancestral and tip states}
  }
}
\description{
Function to generate an annotated edge matrix with ancestral and tip states
}
