% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_algorithm.R
\name{nearest_neighbor_algorithm}
\alias{nearest_neighbor_algorithm}
\title{Identify the phylogenetic nearest neighbor of a tip}
\usage{
nearest_neighbor_algorithm(
  isolate,
  phylogenetic_distance,
  variant_distance,
  metadata,
  annotate = NULL,
  variables_of_interest = NULL,
  comparison = NULL,
  comparison_feature = NULL
)
}
\arguments{
\item{isolate}{Isolate of interest}

\item{phylogenetic_distance}{Matrix of phylogenetic distance}

\item{variant_distance}{Provide a secondary distance measure}

\item{metadata}{Dataframe with trait and variables of interest}

\item{annotate}{Boolean (TRUE/FALSE) indicating if annotation is necessary}

\item{variables_of_interest}{Variables found in the metadata}

\item{comparison}{Boolean (TRUE/FALSE) indicating whether to focus on isolates with/without a trait}

\item{comparison_feature}{Name of variable to compare.}
}
\value{
The nearest neighbor of interest.
}
\description{
This function uses phylogenetics to identify an isolate's nearest neighbor.
}
