% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloaware_regression.R
\name{phyloaware_regression}
\alias{phyloaware_regression}
\title{phyloaware_regression: Phylogenetically-aware regression}
\usage{
phyloaware_regression(
  trait,
  variables,
  df,
  first_present = NULL,
  patient_id = NULL,
  culture_date = NULL,
  multivariable = NULL,
  stepwise_direction = NULL,
  entry_criteria = NULL,
  retention_criteria = NULL,
  confounding_criteria = NULL
)
}
\arguments{
\item{trait}{Trait of interest}

\item{variables}{Exposure variables of interest}

\item{df}{Dataframe}

\item{first_present}{Boolean (i.e., TRUE, FALSE) whether to select a participant's first isolate with the trait.}

\item{patient_id}{Patient identifier variable stored in dataframe df. Required if first_present == TRUE.}

\item{culture_date}{Culture date used to select the participant's first isolate. Must be stored as a variable in the dataframe df. Required if first_present == TRUE.}

\item{multivariable}{Defines the multivariable selection strategy. Options include: 'purposeful', 'AIC', 'pvalue', and 'multivariable.'}

\item{stepwise_direction}{Direction of stepwise selection. Options include: 'both', 'backward', or 'forward'.}

\item{entry_criteria}{P-value for defining candidate variables for multivariable regression. Used in pvalue and purposeful selection)}

\item{retention_criteria}{P-value for retaining candidate variables in model (used in pvalue and purposeful selection)}

\item{confounding_criteria}{Impact on effect size for purposeful selection (used in purposeful selection)}
}
\value{
List with univariable and multivariable results (if requested)
}
\description{
This function performs regression on our dataset
}
