% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloaware_regression.R
\name{phyloaware_regression}
\alias{phyloaware_regression}
\title{phyloaware_regression: Phylogenetically-aware regression}
\usage{
phyloaware_regression(
  trait,
  variables,
  df,
  first_present = NULL,
  patient_id = NULL,
  culture_date = NULL,
  multivariable = NULL,
  stepwise_direction = NULL,
  entry_criteria = NULL,
  retention_criteria = NULL,
  confounding_criteria = NULL
)
}
\arguments{
\item{trait}{Trait of interest}

\item{variables}{Exposure variables of interest}

\item{df}{Dataframe}

\item{first_present}{Whether to identify first present isolate from a patient}

\item{patient_id}{Patient identifier variable stored in dataframe df}

\item{culture_date}{Culture date variable stored in dataframe df}

\item{multivariable}{Boolean (i.e., TRUE, FALSE) whether to perform multivariable regression}

\item{stepwise_direction}{Direction if stepwise multivariable regression is chosen}

\item{entry_criteria}{P-value for defining candidate variables for multivariable regression (used in pvalue and purposeful selection)}

\item{retention_criteria}{P-value for retaining candidate variables in model (used in pvalue and purposeful selection)}

\item{confounding_criteria}{Impact on effect size for purposeful selection (used in purposeful selection)}
}
\value{
List with univariable and multivariable results (if requested)
}
\description{
This function performs regression on our dataset
}
