% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous_permutation_test.R
\name{synchronous_permutation_test}
\alias{synchronous_permutation_test}
\title{synchronous_permutation_test: Synchronous permutation test}
\usage{
synchronous_permutation_test(
  genotype,
  df,
  tr,
  tip_name_var,
  pheno,
  node_states = "joint",
  num_permutations = 1000,
  num_cores = 6
)
}
\arguments{
\item{tr}{Phylogenetic tree}

\item{node_states}{Joint or marginal reconstruction}

\item{comparitor_parent_child_df}{Parent child dataset for a comparitor trait, such as a genotype}

\item{trait_parent_child_df}{Parent child dataset for a trait of interest, such as a phenotype}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}
}
\value{
Synchronous gain and loss events of two traits with p-value permutation testing results
}
\description{
Permutation test for synchronous detection function
}
