% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous_permutation_test.R
\name{synchronous_permutation_test}
\alias{synchronous_permutation_test}
\title{synchronous_permutation_test: Synchronous permutation test}
\usage{
synchronous_permutation_test(
  comparitor,
  df,
  tr,
  tip_name_var,
  trait,
  node_states = "joint",
  confidence_threshold = NULL,
  confidence = NULL,
  num_permutations = 1000,
  num_cores = 1
)
}
\arguments{
\item{comparitor}{Comparitor, such as a genotype or a different trait/phenotype}

\item{df}{Dataframe with comparitor, trait, and tip_name_var}

\item{tr}{Phylogenetic tree}

\item{tip_name_var}{Tip name variable}

\item{trait}{Trait of interest, such as a trait/phenotype}

\item{node_states}{Joint or marginal reconstruction}

\item{confidence_threshold}{Confidence threshold if using marginal reconstruction}

\item{confidence}{Whether to use high or low confidence transition nodes when node_states are marginal}

\item{num_permutations}{Number of permutations. Default: 1000}

\item{num_cores}{Number of cores. Default: 1}
}
\value{
Synchronous gain and loss events of two traits with p-value permutation testing results
}
\description{
Permutation test for synchronous detection function
}
